refactor: decompose session summary into components and hooks
Extract from app/session/summary/[id].tsx (1116→256 lines):
- hooks/useSummaryData.ts — data loading and computed metrics
- hooks/useSummaryActions.ts — rating, notes, template, share actions
- lib/session-display.ts — durationSpoken, delta, deltaTime utilities
- components/session/summary/AchievementsCard.tsx
- components/session/summary/PRsCard.tsx
- components/session/summary/WeightIncreasesCard.tsx
- components/session/summary/ComparisonCard.tsx
- components/session/summary/SetsCard.tsx
- components/session/summary/SummaryFooter.tsx
- hooks/useThemeColors.ts — export ThemeColors type
