+++ fdroid/metadata/com.anomalyco.fitforge.yml
@@ -27,5 +27,5 @@ Description: |
   - CSV export and sharing
   - Dark mode support
 
-CurrentVersion: 0.9.1
-CurrentVersionCode: 14
+CurrentVersion: 0.9.2
+CurrentVersionCode: 15
