+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fitforge",
-  "version": "0.9.1",
+  "version": "0.9.2",
   "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
